changed from Float to BigDecimal for prices added artist popularity

+++ lib/sevendigital/digestion_tract/price_digestor.rb
@@ -14,9 +14,9 @@ module Sevendigital
         price.currency_code = get_required_attribute(c, "code") {|v| v.upcase.to_sym}
       end
       price.currency_symbol = get_required_value(xml_node, "currency")
-      price.value = get_required_value(xml_node,"value") {|v| v.to_f unless v.empty?}
+      price.value = get_required_value(xml_node,"value") {|v| v.to_d unless v.empty?}
       price.formatted_price = get_required_value(xml_node,"formattedPrice")
-      price.rrp = get_optional_value(xml_node,"rrp") {|v| v.to_f unless v.empty?}
+      price.rrp = get_optional_value(xml_node,"rrp") {|v| v.to_d unless v.empty?}
       price.formatted_rrp = get_optional_value(xml_node,"formattedRrp")
       price.on_sale = get_optional_value(xml_node,"onSale") {|v| v.downcase == "true"}
